data-generator/requirements.txt
faker>=22.0.0
pandas>=2.1.0
numpy>=1.26.0
psycopg2-binary>=2.9.9
